Refactor COM traits by organizing modules and replacing string copy utilities with safer pointer handling
