+++ opc_da/src/com/item.rs
@@ -6,7 +6,7 @@ use windows::Win32::Foundation::S_OK;
 use super::{
     base::{Node, Variant},
     bindings::tagOPCITEMPROPERTY,
-    utils::copy_to_com_string,
+    utils::{PointerWriter, TryWriteTo},
 };
 
 pub struct Item {
@@ -42,17 +42,19 @@ impl Node {
     }
 }
 
-impl From<ItemProperty> for tagOPCITEMPROPERTY {
-    fn from(val: ItemProperty) -> Self {
-        tagOPCITEMPROPERTY {
+impl TryFrom<ItemProperty> for tagOPCITEMPROPERTY {
+    type Error = windows_core::Error;
+
+    fn try_from(val: ItemProperty) -> Result<Self, Self::Error> {
+        Ok(tagOPCITEMPROPERTY {
             vtDataType: val.value.get_data_type(),
             wReserved: 0,
             dwPropertyID: val.id,
-            szItemID: copy_to_com_string(&val.name),
-            szDescription: copy_to_com_string(&val.description),
+            szItemID: PointerWriter::try_write_to(&val.name)?,
+            szDescription: PointerWriter::try_write_to(&val.description)?,
             vValue: ManuallyDrop::new(val.value.into()),
             hrErrorID: S_OK,
             dwReserved: 0,
-        }
+        })
     }
 }
