+++ opc_da/src/traits/mod.rs
@@ -1,5 +1,9 @@
+mod def;
+mod group;
 mod item;
 mod server;
 
+pub use def::*;
+pub use group::*;
 pub use item::*;
 pub use server::*;
