+++ opc_da/src/traits/server.rs
@@ -1,4 +1,4 @@
-use crate::com::base::{SystemTime, Variant};
+use super::def::*;
 
 /// Server trait
 ///
@@ -243,137 +243,4 @@ pub trait ServerTrait {
         scope: EnumScope,
         reference_interface_id: Option<u128>,
     ) -> windows_core::Result<windows_core::IUnknown>;
-}
-
-pub struct AvailableProperty {
-    pub property_id: u32,
-    pub description: String,
-    pub data_type: u16,
-}
-
-pub struct ItemPropertyData {
-    pub property_id: u32,
-    pub data: Variant,
-    pub error: windows_core::HRESULT,
-}
-
-pub struct NewItem {
-    pub new_item_id: String,
-    pub error: windows_core::HRESULT,
-}
-
-pub struct ItemProperties {
-    pub error_id: windows_core::HRESULT,
-    pub item_properties: Vec<ItemProperty>,
-}
-
-pub struct ItemProperty {
-    pub data_type: u16,
-    pub property_id: u32,
-    pub item_id: String,
-    pub description: String,
-    pub value: Variant,
-    pub error_id: windows_core::HRESULT,
-}
-
-pub enum BrowseFilter {
-    All,
-    Branches,
-    Items,
-}
-
-pub struct BrowseResult {
-    pub more_elements: bool,
-    pub continuation_point: Option<String>,
-    pub elements: Vec<BrowseElement>,
-}
-
-pub struct BrowseElement {
-    pub name: String,
-    pub item_id: String,
-    pub flag_value: u32,
-    pub item_properties: ItemProperties,
-}
-
-pub enum NamespaceType {
-    Flat,
-    Hierarchical,
-}
-
-pub enum BrowseDirection {
-    Up,
-    Down,
-    To(String),
-}
-
-pub enum BrowseType {
-    Branch,
-    Leaf,
-    Flat,
-}
-
-pub struct ItemWithMaxAge {
-    pub item_id: String,
-    pub max_age: u32,
-}
-
-pub struct Vqt {
-    pub value: Variant,
-    pub quality: u16,
-    pub timestamp: SystemTime,
-}
-
-pub struct OptionalVqt {
-    pub value: Variant,
-    pub quality: Option<u16>,
-    pub timestamp: Option<SystemTime>,
-}
-
-pub struct VqtWithError {
-    pub value: Variant,
-    pub quality: u16,
-    pub timestamp: SystemTime,
-    pub error: windows_core::HRESULT,
-}
-
-pub struct ItemOptionalVqt {
-    pub item_id: String,
-    pub optional_vqt: OptionalVqt,
-}
-
-pub struct GroupInfo {
-    pub server_group: u32,
-    pub revised_update_rate: u32,
-    pub unknown: windows_core::IUnknown,
-}
-
-pub struct ServerStatus {
-    pub start_time: SystemTime,
-    pub current_time: SystemTime,
-    pub last_update_time: SystemTime,
-    pub server_state: ServerState,
-    pub group_count: u32,
-    pub band_width: u32,
-    pub major_version: u16,
-    pub minor_version: u16,
-    pub build_number: u16,
-    pub vendor_info: String,
-}
-
-pub enum ServerState {
-    Running,
-    Failed,
-    NoConfig,
-    Suspended,
-    Test,
-    CommunicationFault,
-}
-
-pub enum EnumScope {
-    PrivateConnections,
-    PublicConnections,
-    AllConnections,
-    Public,
-    Private,
-    All,
 }
